requirements.txt
streamlit
torch
torchvision
facenet-pytorch
Pillow
numpy
pandas
gdown